Add goto output folder and relaunch support to analyses search results.
