+++ src/components/analyses/utils.js
@@ -148,6 +148,33 @@ const getListingPath = (
 ) =>
     `/${NavigationConstants.ANALYSES}?selectedOrder=${order}&selectedOrderBy=${orderBy}&selectedPage=${page}&selectedRowsPerPage=${rowsPerPage}&selectedPermFilter=${permFilter}&selectedTypeFilter=${appTypeFilter}`;
 
+/**
+ * Open a new tab / window with VICE access url is clicked
+ * @param {*} url 
+ */    
+const openInteractiveUrl = (url) => {
+    window.open(url, "_blank");
+};
+
+/**
+ * A custom hook to get `href` and `as` for using with analysis relaunch next/link
+ * 
+ * @param {object} analysis - selected analysis for relaunch
+ * 
+ * @return {array} href and as to be used in next/link
+ */
+const useRelaunchLink = (analysis) => {
+    const href = `/${NavigationConstants.ANALYSES}/[analysisId]/relaunch`;
+    const as = `/${NavigationConstants.ANALYSES}/${analysis?.id}/relaunch`;
+    return [href, as];
+}
+
+const useGotoOutputFolderLink = (analysis) => {
+    const href = `/${NavigationConstants.DATA}/${constants.DATA_STORE_STORAGE_ID}`;
+    const as = `/${NavigationConstants.DATA}/${constants.DATA_STORE_STORAGE_ID}${analysis.resultfolderid}`;
+    return [href, as];
+}
+
 export {
     getAnalysisUser,
     isInteractive,
@@ -159,4 +186,7 @@ export {
     allowAnalysisEdit,
     getAnalysisRelaunchPage,
     getListingPath,
+    openInteractiveUrl,
+    useRelaunchLink,
+    useGotoOutputFolderLink
 };
