--- src/components/search/detailed/styles.js
@@ -1,9 +1,0 @@
-export default (theme) => ({
-    dataLink: {
-        cursor: "pointer",
-        "&:hover": {
-            textDecoration: "underline",
-            color: theme.palette.primary.main,
-        },
-    },
-});
